Fix undo/redo tests.

+++ app/services/editor-commands/commands/remove-item.ts
@@ -79,32 +79,40 @@ export class RemoveItemCommand extends Command {
 
     // If the scene has vertical nodes, remove the corresponding vertical node
     if (this.dualOutputService.views.hasNodeMap()) {
-      this.dualOutputNodeDisplay = this.dualOutputService.views.getVerticalNodeId(this.sceneItemId)
+      this.dualOutputNodeDisplay = this.dualOutputService.views.getVerticalNodeId(
+        this.sceneItemId,
+        this.sceneId,
+      )
         ? 'horizontal'
         : 'vertical';
 
       this.dualOutputNodeId =
         this.dualOutputNodeDisplay === 'horizontal'
-          ? this.dualOutputService.views.getVerticalNodeId(this.sceneItemId)
+          ? this.dualOutputService.views.getVerticalNodeId(this.sceneItemId, this.sceneId)
           : this.dualOutputService.views.getHorizontalNodeId(this.sceneItemId, this.sceneId);
 
       const dualOutputSceneItem = this.scenesService.views.getSceneItem(this.dualOutputNodeId);
-      this.dualOutputNodeSettings = dualOutputSceneItem.getSettings();
 
-      this.reorderDualOutputNodesSubcommand = new ReorderNodesCommand(
-        scene.getSelection(this.dualOutputNodeId),
-        void 0,
-        EPlaceType.After,
-      );
+      // when a scene uses this command, it iterates over both the horizontal and vertical ones
+      // so vertical nodes may have already been deleted
+      if (dualOutputSceneItem) {
+        this.dualOutputNodeSettings = dualOutputSceneItem.getSettings();
 
-      this.reorderDualOutputNodesSubcommand.execute();
+        this.reorderDualOutputNodesSubcommand = new ReorderNodesCommand(
+          scene.getSelection(this.dualOutputNodeId),
+          void 0,
+          EPlaceType.After,
+        );
 
-      dualOutputSceneItem.remove();
+        this.reorderDualOutputNodesSubcommand.execute();
 
-      const nodeToRemoveId =
-        this.dualOutputNodeDisplay === 'horizontal' ? this.sceneItemId : this.dualOutputNodeId;
+        dualOutputSceneItem.remove();
 
-      this.sceneCollectionsService.removeVerticalNode(nodeToRemoveId, this.sceneId);
+        const nodeToRemoveId =
+          this.dualOutputNodeDisplay === 'horizontal' ? this.sceneItemId : this.dualOutputNodeId;
+
+        this.sceneCollectionsService.removeVerticalNode(nodeToRemoveId, this.sceneId);
+      }
     }
 
     // If this was the last item using this source, the underlying source
@@ -128,7 +136,8 @@ export class RemoveItemCommand extends Command {
 
     const scene = this.scenesService.views.getScene(this.sceneId);
 
-    if (this.dualOutputNodeId) {
+    // dual output node settings will only be set if the node was found when removing items
+    if (this.dualOutputNodeSettings) {
       // if the scene has vertical node items, restore both the horizontal and vertical nodes
       const horizontalNodeId =
         this.dualOutputNodeDisplay === 'horizontal' ? this.sceneItemId : this.dualOutputNodeId;
@@ -176,7 +185,15 @@ export class RemoveItemCommand extends Command {
       const item = scene.addSource(this.sourceId, { id: this.sceneItemId, select: false });
 
       this.reorderNodesSubcommand.rollback();
-      item.setSettings(this.settings);
+
+      console.log('setting node settings for sceneItem ', item.id);
+      console.log('this.settings ', this.settings);
+      const horizontalContext = this.videoSettingsService.contexts.horizontal;
+      item.setSettings({
+        ...this.settings,
+        output: horizontalContext,
+        display: 'horizontal',
+      });
     }
   }
 }
